+++ CHANGELOG.md
@@ -32,6 +32,7 @@ All notable changes to this project will be documented in this file. The format 
   - POST `/users` endpoint response from StatusDTO to UserBasicDto (Michał Plebański)
   - `/jurisdictions` endpoint and DB schema ([#1391](https://github.com/bloom-housing/bloom/pull/1391))
   - `/reservedCommunityTypes` endpoint and DB schema ([#1395](https://github.com/bloom-housing/bloom/pull/1395))
+  - `image` field to `listing` model ([#1413](https://github.com/bloom-housing/bloom/pull/1413))
   - reserved_community_type table seeds (`senior` and `specialNeeds`)
 
 - Changed: